# frontend/src/api/openapi.merged.yaml
openapi: 3.1.0
info:
  title: ApeRAG API
  description: ApeRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /quotas:
    get:
      summary: Get user quotas
      description: Get quota information for the current user or all users (admin only)
      tags:
        - quotas
      parameters:
        - name: user_id
          in: query
          description: User ID to get quotas for (admin only, defaults to current user)
          required: false
          schema:
            type: string
            example: user123
        - name: all_users
          in: query
          description: Get quotas for all users (admin only)
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Quota information retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/userQuotaInfo'
                  - $ref: '#/components/schemas/userQuotaList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required for other users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /quotas/{user_id}:
    put:
      summary: Update user quota
      description: Update quota limit for a specific user (admin only)
      tags:
        - quotas
      parameters:
        - name: user_id
          in: path
          description: User ID to update quota for
          required: true
          schema:
            type: string
            example: user123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/quotaUpdateRequest'
      responses:
        '200':
          description: Quota updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/quotaUpdateResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /quotas/{user_id}/recalculate:
    post:
      summary: Recalculate user quota usage
      description: Recalculate and update current usage for all quota types for a user (admin only)
      tags:
        - quotas
      parameters:
        - name: user_id
          in: path
          description: User ID to recalculate quotas for
          required: true
          schema:
            type: string
            example: user123
      responses:
        '200':
          description: Quota usage recalculated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Quota usage recalculated successfully
                  updated_usage:
                    type: object
                    additionalProperties:
                      type: integer
                    example:
                      max_collection_count: 3
                      max_document_count: 25
                      max_bot_count: 2
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    quotaInfo:
      type: object
      description: Quota information for a specific quota type
      properties:
        quota_type:
          type: string
          description: Type of quota
          example: max_collection_count
        quota_limit:
          type: integer
          description: Maximum allowed usage
          example: 10
        current_usage:
          type: integer
          description: Current usage count
          example: 3
        remaining:
          type: integer
          description: Remaining quota available
          example: 7
      required:
        - quota_type
        - quota_limit
        - current_usage
        - remaining
    userQuotaInfo:
      type: object
      description: Complete quota information for a user
      properties:
        user_id:
          type: string
          description: User ID
          example: user123
        username:
          type: string
          description: Username
          example: john_doe
        email:
          type: string
          description: User email
          example: john@example.com
          nullable: true
        role:
          type: string
          description: User role
          example: rw
        quotas:
          type: array
          description: List of quota information
          items:
            $ref: '#/components/schemas/quotaInfo'
      required:
        - user_id
        - username
        - role
        - quotas
    userQuotaList:
      type: object
      description: List of user quota information (admin view)
      properties:
        items:
          type: array
          description: List of user quota information
          items:
            $ref: '#/components/schemas/userQuotaInfo'
      required:
        - items
    quotaUpdateRequest:
      type: object
      description: Request to update user quota
      properties:
        quota_type:
          type: string
          description: Type of quota to update
          example: max_collection_count
          enum:
            - max_collection_count
            - max_document_count
            - max_document_count_per_collection
            - max_bot_count
        new_limit:
          type: integer
          description: New quota limit
          example: 20
          minimum: 0
      required:
        - quota_type
        - new_limit
    quotaUpdateResponse:
      type: object
      description: Response after updating user quota
      properties:
        success:
          type: boolean
          description: Whether the update was successful
          example: true
        message:
          type: string
          description: Status message
          example: Quota updated successfully
        user_id:
          type: string
          description: User ID that was updated
          example: user123
        quota_type:
          type: string
          description: Type of quota that was updated
          example: max_collection_count
        old_limit:
          type: integer
          description: Previous quota limit
          example: 10
        new_limit:
          type: integer
          description: New quota limit
          example: 20
      required:
        - success
        - message
        - user_id
        - quota_type
        - old_limit
        - new_limit
...